+++ src/api/obras.js
@@ -1,0 +1,8 @@
+import request from '@/utils/request'
+
+export function getObras() {
+    return request({
+      url: '/api/v1/camino/',
+      method: 'get'
+    })
+}