+++ src/api/user.js
@@ -1,0 +1,24 @@
+import request from '@/utils/request'
+
+export function login(data) {
+  return request({
+    url: '/login',
+    method: 'post',
+    data
+  })
+}
+
+export function refreshToken(data){
+  return request ({
+    url: '/refreshToken',
+    method: 'post',
+    data
+  })
+}
+
+export function logout() {
+  return request({
+    url: '/user/logout',
+    method: 'post'
+  })
+}