+++ src/router/index.js
@@ -1,27 +1,52 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
+import BusquedaCaminos from '../views/BusquedaCaminos.vue'
 import AltaCamino from '../views/AltaCaminosWork.vue'
 import AnalysisWork from '../components/AnalysisWork.vue'
+import Login from '../views/Login.vue'
+import store from "../store/index";
+import GridResultObra from "../views/GridResultBusqueda.vue"
 
 Vue.use(VueRouter)
 
 const routes = [
   {
+    path: "/login",
+    name: "Login",
+    component: Login,
+    meta: { guest: true },
+  },  
+  {
     path: '/',
     name: 'Home',
-    component: Home
+    //redirect: { name: 'BusquedaCaminos' },
+    redirect: { name: 'Login' },
+    meta: { requiresAuth: false }
+  },
+  {
+    path: '/busqueda-obras',
+    name: 'BusquedaCaminos',
+    component: BusquedaCaminos,
+    meta: { requiresAuth: false }
   },
   {
     path: '/altacamino',
     name: 'AltaCamino',
-    component: AltaCamino
+    component: AltaCamino,
+    meta: { requiresAuth: false }
   },  
   {
     path: '/analisis-de-obra',
     name: 'analisisdeobra',
-    component: AnalysisWork
-  }
+    component: AnalysisWork,
+    meta: { requiresAuth: false }
+  },
+  {
+    path: '/obras',
+    name: 'Obras',
+    component: GridResultObra,
+    meta: { requiresAuth: false }
+  }    
 ]
 
 const router = new VueRouter({
@@ -29,5 +54,20 @@ const router = new VueRouter({
   base: process.env.BASE_URL,
   routes
 })
+
+router.beforeEach((to, from, next) => {
+  const requiresAuth = to.matched.some(record => record.meta.requiresAuth)
+  const isAuthenticated = store.state.user.isAuthenticated
+  console.log(requiresAuth)
+  console.log(isAuthenticated)
+  if (requiresAuth && !isAuthenticated) {
+    console.log('login')
+    next('/login')
+  } else {
+    console.log('else')
+    next()
+  }
+})
+
 
 export default router