+++ src/utils/request.js
@@ -1,9 +1,52 @@
 import axios from 'axios'
+import store from '@/store/'
+import createAuthRefreshInterceptor from "axios-auth-refresh";
 
 const service = axios.create({
-  baseURL: 'http://ccr-back.apps.lab.sct.gob.mx/', 
+  //baseURL: 'http://localhost:4000/', 
+  baseURL:'http://ccr-back.apps.lab.sct.gob.mx/', 
   timeout: 5000, // request timeou
 })
+
+//refresh
+const refreshAuthLogic = failedRequest => service.post('/token', {
+  token: getRefreshToken()
+}).then(tokenRefreshResponse => {
+  console.log('tokenRefreshResponse')
+  console.log(tokenRefreshResponse)
+  store.dispatch('user/setAccessToken', tokenRefreshResponse.accessToken)
+  failedRequest.response.config.headers["Authorization"] = 'Bearer ' + tokenRefreshResponse.accessToken
+  return Promise.resolve()
+});
+
+const statusCodes = {
+  statusCodes: [ 401, 403 ] // default: [ 401 ]
+}
+createAuthRefreshInterceptor(service, refreshAuthLogic, statusCodes);
+
+
+function getRefreshToken(){
+  console.log('refreshhhhhhhhh')
+  const camino = JSON.parse(localStorage.getItem('camino'))
+  if(camino) return camino.user.refreshToken
+  return ''
+}
+
+// request interceptor
+function getAccessToken() {
+  console.log('gettttt')
+  console.log(JSON.parse(JSON.stringify(localStorage.getItem('camino'))))
+  const camino = JSON.parse(localStorage.getItem('camino'))
+  console.log(camino)
+  if(camino) return camino.user.token
+  return ''
+}
+
+service.interceptors.request.use(request => {
+  request.headers['Authorization'] = "Bearer " + getAccessToken();
+  return request;
+});
+
 
 // response interceptor
 service.interceptors.response.use(